refactor(templates): extraire organisms settings + templates pages complexes (project-detail, user-settings, accept-invitation)
